@@ -1,6 +1,7 @@
-﻿using Lib.SQL.Adapter;
-using Lib.SQL.QueryBuilder.Operator;
-using Lib.SQL.QueryBuilder.Sequences;
+﻿using System.Data.SQLite;
+using Lib.SQL.Adapter;
+using Lib.SQL.Operation.QueryBuilder.Operator;
+using Lib.SQL.Operation.QueryBuilder.Sequences;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Lib.SQL.SQLite.Test
@@ -8,21 +9,21 @@ namespace Lib.SQL.SQLite.Test
     [TestClass]
     public class TestDbal
     {
-        private readonly DbAdapter _adapter =  Adapter.CreateFromPlainScript(":memory:", "CREATE TABLE example(colA TEXT, colB TEXT)", true);
+        private readonly DbAdapter _adapter =  Adapter.CreateFromPlainScript(new SQLiteConnectionStringBuilder { DataSource = ":memory:" }, "CREATE TABLE example(colA TEXT, colB TEXT)", true);
         private readonly Table _table;
 
         public TestDbal()
         {
-            _table = new Table(() => _adapter, "example", "colA", "colB");
+            _table = new Table("example", "colA", "colB");
         }
 
         [TestMethod]
         public void TestInsert()
         {
-            var count = _table.Insert().Values("a", "b").Values("c", "d").Execute();
+            var count = _table.Insert().Values("a", "b").Values("c", "d").ExecuteOn(_adapter);
             Assert.AreEqual(2, count);
 
-            var selected = _table.SelectLines().OrderBy("colA", OrderDirection.Asc).Execute();
+            var selected = _table.SelectLines().OrderBy("colA", OrderDirection.Asc).ExecuteOn(_adapter);
             Assert.AreEqual(2, selected.Count);
             Assert.AreEqual("a", selected[0]["colA"].ToString());
             Assert.AreEqual("b", selected[0]["colB"].ToString());
@@ -33,10 +34,10 @@ namespace Lib.SQL.SQLite.Test
         [TestMethod]
         public void TestUpdate()
         {
-            _table.Insert().Values("a", "b").Values("c", "d").Execute();
-            _table.Update().Set("colA", "z").Where("colB", Is.EqualWith, "b").Execute();
+            _table.Insert().Values("a", "b").Values("c", "d").ExecuteOn(_adapter);
+            _table.Update().Set("colA", "z").Where("colB", Is.EqualWith, "b").ExecuteOn(_adapter);
 
-            var selected = _table.SelectLines().OrderBy("colB", OrderDirection.Asc).Execute();
+            var selected = _table.SelectLines().OrderBy("colB", OrderDirection.Asc).ExecuteOn(_adapter);
             Assert.AreEqual(2, selected.Count);
             Assert.AreEqual("z", selected[0]["colA"].ToString());
             Assert.AreEqual("b", selected[0]["colB"].ToString());
@@ -47,10 +48,10 @@ namespace Lib.SQL.SQLite.Test
         [TestMethod]
         public void TestDelete()
         {
-            _table.Insert().Values("a", "b").Values("c", "d").Execute();
-            _table.Delete().Where("colB", Is.EqualWith, "b").Execute();
+            _table.Insert().Values("a", "b").Values("c", "d").ExecuteOn(_adapter);
+            _table.Delete().Where("colB", Is.EqualWith, "b").ExecuteOn(_adapter);
 
-            var selected = _table.SelectLines().Execute();
+            var selected = _table.SelectLines().ExecuteOn(_adapter);
             Assert.AreEqual(1, selected.Count);
             Assert.AreEqual("c", selected[0]["colA"].ToString());
             Assert.AreEqual("d", selected[0]["colB"].ToString());
@@ -59,30 +60,30 @@ namespace Lib.SQL.SQLite.Test
         [TestMethod]
         public void TestSelect()
         {
-            var rowId = _table.Insert().Values("a", "b").Values("c", "d").ExecuteAndReturnRowId();
-            Assert.AreEqual(rowId, _table.Select("last_insert_rowid()").Execute());
+            var rowId = _table.Insert().Values("a", "b").Values("c", "d").ExecuteOnAndReturnRowId(_adapter);
+            Assert.AreEqual(rowId, _table.Select("last_insert_rowid()").ExecuteOn(_adapter));
         }
 
         [TestMethod]
         public void TestExists()
         {
-            var selected = _table.Exists().Execute();
+            var selected = _table.Exists().ExecuteOn(_adapter);
             Assert.AreEqual(false, selected);
 
-            _table.Insert().Values("a", "b").Values("c", "d").Execute();
-            selected = _table.Exists().Where("colB", Is.EqualWith, "b").Execute();
+            _table.Insert().Values("a", "b").Values("c", "d").ExecuteOn(_adapter);
+            selected = _table.Exists().Where("colB", Is.EqualWith, "b").ExecuteOn(_adapter);
             Assert.AreEqual(true, selected);
         }
 
         [TestMethod]
         public void TestLastInsertedId()
         {
-            var notInserted = _table.LastInsertedId;
+            var notInserted = _adapter.LastInsertedId;
             Assert.AreEqual(0, notInserted);
 
-            var id = _table.Insert().Values("a", "b").ExecuteAndReturnRowId();
+            var id = _table.Insert().Values("a", "b").ExecuteOnAndReturnRowId(_adapter);
 
-            var lastInserted = _table.LastInsertedId;
+            var lastInserted = _adapter.LastInsertedId;
             Assert.AreEqual(id, lastInserted);
         }
     }
