@@ -1,6 +1,6 @@
 ﻿using System;
-using Lib.SQL.QueryBuilder;
-using Lib.SQL.QueryBuilder.Operator;
+using Lib.SQL.Operation.QueryBuilder;
+using Lib.SQL.Operation.QueryBuilder.Operator;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Lib.SQL.Test.QueryBuilder
