@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using Lib.SQL.QueryBuilder;
-using Lib.SQL.QueryBuilder.Sequences;
+using Lib.SQL.Operation.QueryBuilder;
+using Lib.SQL.Operation.QueryBuilder.Sequences;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Lib.SQL.Test.QueryBuilder
