@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 using Lib.SQL.Adapter;
 using Lib.SQL.Executor.Collections;
 
@@ -6,9 +7,10 @@ namespace Lib.SQL.Executor
 {
     public class MultipleLinesExecutor : ExecutorAbstract<ResultLines>
     {
-        protected override ResultLines ExecuteOnAdapter(DbAdapter adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null)
-        {
-            return new ResultLines(adapter.FetchLines(sql, parameters));
-        }
+        protected override ResultLines ExecuteOnAdapter(ICommandChannel adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null) 
+            => new ResultLines(adapter.FetchLines(sql, parameters));
+
+        protected override async Task<ResultLines> ExecuteOnAdapterAsync(ICommandChannel adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null)
+            => new ResultLines(await adapter.FetchLinesAsync(sql, parameters));
     }
 }
