@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
-using System.IO;
 using System.Linq;
+using System.Threading.Tasks;
 using Lib.SQL.Adapter;
 using Lib.SQL.Executor.Collections;
 
@@ -8,12 +8,24 @@ namespace Lib.SQL.Executor
 {
     public class SingleColumnExecutor : ExecutorAbstract<object[]>
     {
-        public string Column;
-
-        protected override object[] ExecuteOnAdapter(DbAdapter adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null)
+        protected override object[] ExecuteOnAdapter(ICommandChannel adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null)
         {
-            if(string.IsNullOrWhiteSpace(Column)) throw new InvalidDataException("Définissez le nom de la colonne à récupérer avant de lancer la requête.");
-            return new ResultLines(adapter.FetchLines(sql, parameters)).Select(line => line[Column]).ToArray();
+            var resultLines = new ResultLines(adapter.FetchLines(sql, parameters));
+            return ParseResultLines(resultLines);
+        }
+
+        private static object[] ParseResultLines(ResultLines lines)
+        {
+            var columnName = lines.FirstOrDefault()?.First().Key;
+            return columnName is null 
+                ? new object[0] 
+                : lines.Select(line => line[columnName]).ToArray();
+        }
+
+        protected override async Task<object[]> ExecuteOnAdapterAsync(ICommandChannel adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null)
+        {
+            var lines = new ResultLines(await adapter.FetchLinesAsync(sql, parameters));
+            return ParseResultLines(lines);
         }
     }
 }
