@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 using Lib.SQL.Adapter;
 using Lib.SQL.Executor.Collections;
 
@@ -6,9 +7,10 @@ namespace Lib.SQL.Executor
 {
     public class SingleLineExecutor : ExecutorAbstract<ResultLine>
     {
-        protected override ResultLine ExecuteOnAdapter(DbAdapter adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null)
-        {
-            return new ResultLine(adapter.FetchLine(sql, parameters));
-        }
+        protected override ResultLine ExecuteOnAdapter(ICommandChannel adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null) 
+            => new ResultLine(adapter.FetchLine(sql, parameters));
+
+        protected override async Task<ResultLine> ExecuteOnAdapterAsync(ICommandChannel adapter, string sql, IEnumerable<KeyValuePair<string, object>> parameters = null)
+            => new ResultLine(await adapter.FetchLineAsync(sql, parameters));
     }
 }
