@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using Lib.SQL.Executor;
-using Lib.SQL.QueryBuilder;
-using Lib.SQL.QueryBuilder.Operator;
-using Lib.SQL.QueryBuilder.Sequences;
-using Lib.SQL.QueryBuilder.Sequences.Where;
+using Lib.SQL.Operation.QueryBuilder;
+using Lib.SQL.Operation.QueryBuilder.Operator;
+using Lib.SQL.Operation.QueryBuilder.Sequences;
+using Lib.SQL.Operation.QueryBuilder.Sequences.Where;
 
 namespace Lib.SQL.Operation
 {
